+++ src/main/kotlin/com/github/lanahra/modifierutils/Constants.kt
@@ -1,0 +1,4 @@
+package com.github.lanahra.modifierutils
+
+const val MODIFIER_FINAL = "final"
+const val MODIFIER_STATIC = "static"